Basic audit log (#3)
This is a stripped down version of the access logger found in the
Akka HTTP template: intercepting req/resp at the HTTP level is
problematic cause (i) is an abstraction leakage (IMHO), (ii) does
not offer all the needed info and (iii) most of the machinery is
hidden (e.g. all calls are a POST). For the moment is the best I
could do.

Note that you will always see a 200 OK even if the response was a
(e.g.) 'Bad Request' given that everything gets tunnelled through an
always successful HTTP response.